保存商品

+++ gulimall-product/src/main/java/com/atguigu/gulimall/product/entity/params/Bounds.java
@@ -2,12 +2,14 @@ package com.atguigu.gulimall.product.entity.params;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 /**
  * @author Administrator
  */
 @Data
 public class Bounds {
 
-    private Integer buyBounds;
-    private Integer growBounds;
+    private BigDecimal buyBounds;
+    private BigDecimal growBounds;
 }
